+++ src/preferences/interaction.rs
@@ -5,4 +5,5 @@ use serde::{Deserialize, Serialize};
 pub struct InteractionPreferences {
     pub twist_duration: f32,
     pub dynamic_twist_speed: bool,
+    pub fade_duration: f32,
 }